requirements.txt
# Core web framework
Flask==3.0.0
Flask-CORS==4.0.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4
pydub==0.25.1

# TTS engines (local, no API costs)
piper-tts==1.2.0

# Environment management
python-dotenv==1.0.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Development dependencies
pytest==7.4.3
black==23.12.1

# Optional dependencies (install separately if needed for better pitch detection):
# praat-parselmouth>=0.4.0
# aubio==0.4.9
